+++ CHANGES.rst
@@ -15,7 +15,7 @@ Other changes:
  -
 
 Bugfix:
- -
+ - Fix / Crash when session store is null in event stream #3158
 
 Translations:
  -